Use Ace editor instead of textarea
The Ace editor, http://ace.ajax.org, is a
neat tool for editing code in a browser.

It might be easier to use Ace as a base
if we want to have basic styling in the
notes. More importantly, I'm going to see
if we can make hyperlinks as we type.

After struggling with getting the ui-ace
component to work, I gave up and used
this plunker as a base:

http://plnkr.co/edit/JJe0mf9RmDiSYkfahdnq

As commitment-level is medium at this point,
we're pointing to the ace.js hosted on GitHub
for the time being.