Fix: Init the editor from within the directive
Explicitly calling $apply() within the controller
had a few issues, mainly that we sometimes were
already in the middle of a $digest, plus the
controller shouldn't really need to know about
that stuff.

We fix this by listening to the document change
events within the ace-editor directive, and by
introducing a hack that I can live with for now.