Add authentication to the client
Authorize with the server by posting to /auth.
Check /permissions to see what we can do.

UX:

Hide authcode box and move the focus to the
editor when a correct authcode is entered.

Prevent the menu (which includes the authcode box)
from hiding if we're entering in an authcode.

Only show authcode box when we don't have
write permission.